backend: move Backend + BackendRegistry to commonMain
The commonMain ports of EmberTensor, Loss, Optimizer, and BackendUtils
reference ai.solace.ember.backend.BackendRegistry and the top-level
getBackend/setBackend/isBackendAvailable/getAvailableBackends/
autoSelectBackend facade, but those declarations only existed in
macosArm64Main. That broke commonMain compilation across every other
target with ~45 'Unresolved reference: BackendRegistry' errors.

Split the registry along its real platform axis:

- Backend (the interface) moves to commonMain as-is. It depends only on
  ai.solace.ember.tensor.common.DType which already lives in commonMain.
- BackendRegistry (the object) moves to commonMain as a pure registry:
  registerBackend / getBackend / setBackend / getCurrentBackend /
  getAvailableBackends, plus the top-level getBackend / setBackend /
  isBackendAvailable / getAvailableBackends facade.
- initializePlatformBackends() and autoSelectBackend() become expect
  functions. The macOS actual registers MegaTensorBackend + MetalBackend
  and prefers Metal when available; linuxX64 / linuxArm64 / mingwX64 /
  js actuals ship as no-ops that throw a clear IllegalStateException if
  callers reach autoSelectBackend without having registered a backend
  themselves.

ensureBackendsInitialized() lazily runs the platform actual on first
read from the registry, preserving the old eager-init behavior for
existing callers without coupling commonMain to concrete backends.
